Split Github scripts to always allow panel-based tipping
