+++ scripts/brave_rewards/publisher/github/publisherInfo.ts
@@ -1,0 +1,82 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+import { MediaMetaData } from '../common/types'
+import { port, sendErrorResponse } from './messaging'
+
+import * as types from './types'
+import * as utils from './utils'
+
+const sendForExcludedPage = () => {
+  const url = `https://${types.mediaDomain}`
+  const publisherKey = types.mediaDomain
+  const publisherName = types.mediaDomain
+  const mediaKey = ''
+  const favIconUrl = ''
+
+  if (!port) {
+    return
+  }
+
+  port.postMessage({
+    type: 'SavePublisherVisit',
+    mediaType: '',
+    data: {
+      url,
+      publisherKey,
+      publisherName,
+      mediaKey,
+      favIconUrl
+    }
+  })
+}
+
+const sendForStandardPage = (url: URL) => {
+  const screenName = utils.getScreenNameFromUrl(url)
+  if (!screenName) {
+    sendErrorResponse('Invalid screen name')
+    return
+  }
+
+  return utils.getMediaMetaData(screenName)
+    .then((mediaMetaData: MediaMetaData) => {
+      const userId = mediaMetaData.user.id
+      const publisherKey = utils.buildPublisherKey(userId)
+      const publisherName = mediaMetaData.user.fullName
+      if (!publisherName) {
+        sendErrorResponse('Invalid publisher name')
+        return
+      }
+
+      const mediaKey = ''
+      const favIconUrl = mediaMetaData.user.favIconUrl
+
+      const profileUrl = utils.buildProfileUrl(screenName)
+
+      if (!port) {
+        return
+      }
+
+      port.postMessage({
+        type: 'SavePublisherVisit',
+        mediaType: types.mediaType,
+        data: {
+          url: profileUrl,
+          publisherKey,
+          publisherName,
+          mediaKey,
+          favIconUrl
+        }
+      })
+    })
+}
+
+export const send = () => {
+  const url = new URL(location.href)
+  if (utils.isExcludedPath(url.pathname)) {
+    sendForExcludedPage()
+  } else {
+    sendForStandardPage(url)
+  }
+}
